[WFLY-1923] / [WFLY-490] Where no authentication is required ensure an 'anonymous' user is associated with the current request.
